requirements.txt
pygame
sounddevice
pillow
